调整目录
