+++ app/src/main/AndroidManifest.xml
@@ -18,10 +18,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".module.QuestionPage.QuestionActivity"/>
-        <activity android:name=".module.Comment.CommentActivity"/>
-        <activity android:name=".module.AnwserPage.AnswerActivity"/>
-        <activity android:name=".module.NewPage.NewActivity"/>
+        <activity android:name=".module.Forum.QuestionPage.QuestionActivity"/>
+        <activity android:name=".module.Forum.QuestionPage.Comment.CommentActivity"/>
+        <activity android:name=".module.Forum.QuestionPage.AnwserPage.AnswerActivity"/>
+        <activity android:name=".module.Forum.NewPage.NewActivity"/>
     </application>
 
     <!-- 友盟统计 -->
