+++ app/src/main/java/us/xingkong/jueqian/module/Forum/NewPage/NewPresenter.java
@@ -1,4 +1,4 @@
-package us.xingkong.jueqian.module.NewPage;
+package us.xingkong.jueqian.module.Forum.NewPage;
 
 import us.xingkong.jueqian.base.BasePresenterImpl;
