+++ app/src/main/res/layout/activity_new.xml
@@ -46,4 +46,10 @@
             android:scrollbars="vertical"></EditText>
 
     </LinearLayout>
+    <Switch
+        android:padding="8dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="匿名发布"/>
+
 </LinearLayout>
