+++ app/src/main/res/layout/activity_question.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="us.xingkong.jueqian.module.QuestionPage.QuestionActivity">
+    tools:context=".module.Forum.QuestionPage.QuestionActivity">
 
     <include layout="@layout/view_toolbar" />
